Refactor MaskDINO decoder and encoder modules
